+++ src/connectedk8s/azext_connectedk8s/_params.py
@@ -4,7 +4,9 @@
 # --------------------------------------------------------------------------------------------
 # pylint: disable=line-too-long
 
+from argcomplete.completers import FilesCompleter
 from azure.cli.core.commands.parameters import get_location_type
+from azure.cli.core.commands.parameters import (file_type)
 from azure.cli.core.commands.validators import get_default_location_from_resource_group
 
 
@@ -21,6 +23,7 @@ def load_arguments(self, _):
         c.argument('https_proxy', options_list=['--proxy-https'], help='Https proxy URL to be used.')
         c.argument('http_proxy', options_list=['--proxy-http'], help='Http proxy URL to be used.')
         c.argument('no_proxy', options_list=['--proxy-skip-range'], help='List of URLs/CIDRs for which proxy should not to be used.')
+        c.argument('proxy_cert', options_list=['--proxy-cert'], type=file_type, completer=FilesCompleter(), help='Path to the certificate file for proxy')
 
     with self.argument_context('connectedk8s update') as c:
         c.argument('cluster_name', options_list=['--name', '-n'], id_part='name', help='The name of the connected cluster.')
@@ -29,6 +32,7 @@ def load_arguments(self, _):
         c.argument('https_proxy', options_list=['--proxy-https'], help='Https proxy URL to be used.')
         c.argument('http_proxy', options_list=['--proxy-http'], help='Http proxy URL to be used.')
         c.argument('no_proxy', options_list=['--proxy-skip-range'], help='List of URLs/CIDRs for which proxy should not to be used.')
+        c.argument('proxy_cert', options_list=['--proxy-cert'], type=file_type, completer=FilesCompleter(), help='Path to the certificate file for proxy')
 
     with self.argument_context('connectedk8s list') as c:
         pass