+++ src/storage-blob-preview/azext_storage_blob_preview/_params.py
@@ -6,12 +6,15 @@
 
 from azure.cli.core.commands.validators import validate_tags
 from azure.cli.core.commands.parameters import (file_type, get_enum_type, get_three_state_flag)
-
+from azure.cli.core.local_context import LocalContextAttribute, LocalContextAction
 
 from ._validators import (validate_metadata, get_permission_validator, get_permission_help_string,
                           validate_blob_type, validate_included_datasets_v2, add_progress_callback,
-                          validate_storage_data_plane_list, as_user_validator, blob_tier_validator)
-from .profiles import CUSTOM_DATA_STORAGE_BLOB
+                          validate_storage_data_plane_list, as_user_validator, blob_tier_validator,
+                          validate_container_delete_retention_days, validate_delete_retention_days,
+                          process_resource_group)
+
+from .profiles import CUSTOM_DATA_STORAGE_BLOB, CUSTOM_MGMT_STORAGE
 
 
 def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statements, too-many-lines
@@ -19,8 +22,16 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
 
     from knack.arguments import CLIArgumentType
 
+    from azure.cli.core.commands.parameters import get_resource_name_completion_list
+
     from .sdkutil import get_table_data_type
     from .completers import get_storage_name_completion_list
+
+    acct_name_type = CLIArgumentType(options_list=['--account-name', '-n'], help='The storage account name.',
+                                     id_part='name',
+                                     completer=get_resource_name_completion_list('Microsoft.Storage/storageAccounts'),
+                                     local_context_attribute=LocalContextAttribute(
+                                         name='storage_account_name', actions=[LocalContextAction.GET]))
 
     t_base_blob_service = self.get_sdk('blob.baseblobservice#BaseBlobService')
     t_file_service = self.get_sdk('file#FileService')
@@ -76,7 +87,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
              'where the previous generator stopped.')
 
     num_results_type = CLIArgumentType(
-        default=5000, validator=validate_storage_data_plane_list,
+        default=5000, validator=validate_storage_data_plane_list, options_list='--num-results',
         help='Specify the maximum number to return. If the request does not specify '
         'num_results, or specifies a value greater than 5000, the server will return up to 5000 items. Note that '
         'if the listing operation crosses a partition boundary, then the service will return a continuation token '
@@ -104,6 +115,9 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
     tags_condition_type = CLIArgumentType(
         options_list='--tags-condition', min_api='2019-12-12',
         help='Specify a SQL where clause on blob tags to operate only on blobs with a matching value.')
+    timeout_type = CLIArgumentType(
+        help='Request timeout in seconds. Applies to each call to the service.', type=int
+    )
 
     with self.argument_context('storage') as c:
         c.argument('container_name', container_name_type)
@@ -116,6 +130,52 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
                    help='Metadata in space-separated key=value pairs. This overwrites any existing metadata.',
                    validator=validate_metadata)
         c.argument('timeout', help='Request timeout in seconds. Applies to each call to the service.', type=int)
+
+    with self.argument_context('storage account blob-service-properties show',
+                               resource_type=CUSTOM_MGMT_STORAGE) as c:
+        c.argument('account_name', acct_name_type, id_part=None)
+        c.argument('resource_group_name', required=False, validator=process_resource_group)
+
+    with self.argument_context('storage account blob-service-properties update',
+                               resource_type=CUSTOM_MGMT_STORAGE) as c:
+        c.argument('account_name', acct_name_type, id_part=None)
+        c.argument('resource_group_name', required=False, validator=process_resource_group)
+        c.argument('enable_change_feed', arg_type=get_three_state_flag(), min_api='2019-04-01')
+        c.argument('enable_container_delete_retention',
+                   arg_type=get_three_state_flag(),
+                   options_list=['--enable-container-delete-retention', '--container-retention'],
+                   arg_group='Container Delete Retention Policy', min_api='2019-06-01',
+                   help='Enable container delete retention policy for container soft delete when set to true. '
+                        'Disable container delete retention policy when set to false.')
+        c.argument('container_delete_retention_days',
+                   options_list=['--container-delete-retention-days', '--container-days'],
+                   type=int, arg_group='Container Delete Retention Policy',
+                   min_api='2019-06-01', validator=validate_container_delete_retention_days,
+                   help='Indicate the number of days that the deleted container should be retained. The minimum '
+                        'specified value can be 1 and the maximum value can be 365.')
+        c.argument('enable_delete_retention', arg_type=get_three_state_flag(), arg_group='Delete Retention Policy',
+                   min_api='2018-07-01')
+        c.argument('delete_retention_days', type=int, arg_group='Delete Retention Policy',
+                   validator=validate_delete_retention_days, min_api='2018-07-01')
+        c.argument('enable_restore_policy', arg_type=get_three_state_flag(), arg_group='Restore Policy',
+                   min_api='2019-06-01', help="Enable blob restore policy when it set to true.")
+        c.argument('restore_days', type=int, arg_group='Restore Policy',
+                   min_api='2019-06-01', help="The number of days for the blob can be restored. It should be greater "
+                   "than zero and less than Delete Retention Days.")
+        c.argument('enable_versioning', arg_type=get_three_state_flag(), help='Versioning is enabled if set to true.',
+                   min_api='2019-06-01')
+        c.argument('enable_last_access_tracking', arg_type=get_three_state_flag(), min_api='2019-06-01',
+                   options_list=['--enable-last-access-tracking', '-t'],
+                   help='When set to true last access time based tracking policy is enabled.')
+
+    with self.argument_context('storage account management-policy create') as c:
+        c.argument('policy', type=file_type, completer=FilesCompleter(),
+                   help='The Storage Account ManagementPolicies Rules, in JSON format. See more details in: '
+                        'https://docs.microsoft.com/azure/storage/common/storage-lifecycle-managment-concepts.')
+        c.argument('account_name', help='The name of the storage account within the specified resource group.')
+
+    with self.argument_context('storage account management-policy update') as c:
+        c.argument('account_name', help='The name of the storage account within the specified resource group.')
 
     with self.argument_context('storage blob') as c:
         c.argument('blob_name', options_list=('--name', '-n'), arg_type=blob_name_type)
@@ -291,7 +351,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.argument('num_results', arg_type=num_results_type)
         c.argument('prefix',
                    help='Filter the results to return only blobs whose name begins with the specified prefix.')
-        c.argument('show_next_marker', action='store_true',
+        c.argument('show_next_marker', action='store_true', is_preview=True,
                    help='Show nextMarker in result when specified.')
 
     for item in ['show', 'update']:
@@ -408,3 +468,25 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.ignore('sas_token')
         c.argument('full_uri', action='store_true', is_preview=True,
                    help='Indicate that this command return the full blob URI and the shared access signature token.')
+
+    with self.argument_context('storage container list') as c:
+        c.extra('timeout', timeout_type)
+        c.argument('marker', arg_type=marker_type)
+        c.argument('num_results', arg_type=num_results_type)
+        c.argument('prefix',
+                   help='Filter the results to return only blobs whose name begins with the specified prefix.')
+        c.argument('include_metadata', arg_type=get_three_state_flag(),
+                   help='Specify that container metadata to be returned in the response.')
+        c.argument('show_next_marker', action='store_true', is_preview=True,
+                   help='Show nextMarker in result when specified.')
+        c.argument('include_deleted', arg_type=get_three_state_flag(), min_api='2020-02-10',
+                   help='Specify that deleted containers to be returned in the response. This is for container restore '
+                   'enabled account. The default value is `False`')
+
+    with self.argument_context('storage container restore') as c:
+        c.argument('deleted_container_name', options_list=['--name', '-n'],
+                   help='Specify the name of the deleted container to restore.')
+        c.argument('deleted_container_version', options_list=['--deleted-version'],
+                   help='Specify the version of the deleted container to restore.')
+        c.argument('new_name', help='The new name for the deleted container to be restored to.')
+        c.extra('timeout', timeout_type)