+++ src/vm-repair/azext_vm_repair/custom.py
@@ -4,10 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 # pylint: disable=line-too-long, too-many-locals, too-many-statements, broad-except, too-many-branches
-import json
 import timeit
-import os
-import pkgutil
 import traceback
 import requests
 from knack.log import get_logger
@@ -33,6 +30,7 @@ from .repair_utils import (
     _check_script_succeeded,
     _fetch_disk_info,
     _unlock_singlepass_encrypted_disk,
+    _invoke_run_command
 )
 from .exceptions import AzCommandError, SkuNotAvailableError, UnmanagedDiskCopyError, WindowsOsNotAvailableError, RunScriptNotFoundForIdError
 logger = get_logger(__name__)
@@ -106,7 +104,7 @@ def create(cmd, vm_name, resource_group_name, repair_password=None, repair_usern
 
             # Handle encrypted VM cases
             if unlock_encrypted_vm:
-                _unlock_singlepass_encrypted_disk(source_vm, is_linux, repair_group_name, repair_vm_name)
+                _unlock_singlepass_encrypted_disk(source_vm, resource_group_name, repair_vm_name, repair_group_name, copy_disk_name, is_linux)
 
         # UNMANAGED DISK
         else:
@@ -301,47 +299,36 @@ def run(cmd, vm_name, resource_group_name, run_id=None, repair_vm_id=None, custo
     # Init command helper object
     command = command_helper(logger, cmd, 'vm repair run')
 
-    REPAIR_DIR_NAME = 'azext_vm_repair'
-    SCRIPTS_DIR_NAME = 'scripts'
     LINUX_RUN_SCRIPT_NAME = 'linux-run-driver.sh'
     WINDOWS_RUN_SCRIPT_NAME = 'win-run-driver.ps1'
-    RUN_COMMAND_RUN_SHELL_ID = 'RunShellScript'
-    RUN_COMMAND_RUN_PS_ID = 'RunPowerShellScript'
 
     try:
         # Fetch VM data
         source_vm = get_vm(cmd, resource_group_name, vm_name)
-
-        # Build absoulte path of driver script
-        loader = pkgutil.get_loader(REPAIR_DIR_NAME)
-        mod = loader.load_module(REPAIR_DIR_NAME)
-        rootpath = os.path.dirname(mod.__file__)
         is_linux = _is_linux_os(source_vm)
+
         if is_linux:
-            run_script = os.path.join(rootpath, SCRIPTS_DIR_NAME, LINUX_RUN_SCRIPT_NAME)
-            command_id = RUN_COMMAND_RUN_SHELL_ID
+            script_name = LINUX_RUN_SCRIPT_NAME
         else:
-            run_script = os.path.join(rootpath, SCRIPTS_DIR_NAME, WINDOWS_RUN_SCRIPT_NAME)
-            command_id = RUN_COMMAND_RUN_PS_ID
+            script_name = WINDOWS_RUN_SCRIPT_NAME
 
         # If run_on_repair is False, then repair_vm is the source_vm (scripts run directly on source vm)
         repair_vm_id = parse_resource_id(repair_vm_id)
         repair_vm_name = repair_vm_id['name']
         repair_resource_group = repair_vm_id['resource_group']
 
-        repair_run_command = 'az vm run-command invoke -g {rg} -n {vm} --command-id {command_id} ' \
-                             '--scripts "@{run_script}" -o json' \
-                             .format(rg=repair_resource_group, vm=repair_vm_name, command_id=command_id, run_script=run_script)
+        run_command_params = []
+        additional_scripts = []
 
         # Normal scenario with run id
         if not custom_script_file:
             # Fetch run path from GitHub
             repair_script_path = _fetch_run_script_path(run_id)
-            repair_run_command += ' --parameters script_path="./{repair_script}"'.format(repair_script=repair_script_path)
+            run_command_params.append('script_path="./{}"'.format(repair_script_path))
         # Custom script scenario for script testers
         else:
-            # no-op run id
-            repair_run_command += ' "@{custom_file}" --parameters script_path=no-op'.format(custom_file=custom_script_file)
+            run_command_params.append('script_path=no-op')
+            additional_scripts.append(custom_script_file)
 
         # Append Parameters
         if parameters:
@@ -349,9 +336,7 @@ def run(cmd, vm_name, resource_group_name, run_id=None, repair_vm_id=None, custo
                 param_string = _process_bash_parameters(parameters)
             else:
                 param_string = _process_ps_parameters(parameters)
-            # Work around for run-command bug, unexpected behavior with space characters
-            param_string = param_string.replace(' ', '%20')
-            repair_run_command += ' params="{}"'.format(param_string)
+            run_command_params.append('params="{}"'.format(param_string))
         if run_on_repair:
             vm_string = 'repair VM'
         else:
@@ -360,18 +345,8 @@ def run(cmd, vm_name, resource_group_name, run_id=None, repair_vm_id=None, custo
 
         # Run script and measure script run-time
         script_start_time = timeit.default_timer()
-        return_str = _call_az_command(repair_run_command)
+        stdout, stderr = _invoke_run_command(script_name, repair_vm_name, repair_resource_group, is_linux, run_command_params, additional_scripts)
         command.script.run_time = timeit.default_timer() - script_start_time
-
-        # Extract stdout and stderr, if stderr exists then possible error
-        run_command_return = json.loads(return_str)
-        if is_linux:
-            run_command_message = run_command_return['value'][0]['message'].split('[stdout]')[1].split('[stderr]')
-            stdout = run_command_message[0].strip('\n')
-            stderr = run_command_message[1].strip('\n')
-        else:
-            stdout = run_command_return['value'][0]['message']
-            stderr = run_command_return['value'][1]['message']
         logger.debug("stderr: %s", stderr)
 
         # Parse through stdout to populate log properties: 'level', 'message'