+++ src/spring-cloud/HISTORY.md
@@ -1,5 +1,13 @@
 Release History
 ===============
+1.1.1
+-----
+* Reimport the updated version of Python SDK.
+
+1.1.0
+-----
+* Support Steeltoe feature.
+
 1.0.1
 -----
 * Optimize VNet Injection validator